Redo DMG_NERVEGAS resistance logic
Turns out, I was right when I thought the first way was bad.
When the initial DMG_NERVEGAS damage in BA2_AddInfection was applied with a viral sample, ent:Health would not update to reflect the damage taken. To be fair, there is NO legitimate reason for this to occur -- I'm assuming it's a bug in Garry's Mod. Regardless, it is a bug, and resulted in needing a new solution. THAT SOLUTION IS:

- Create a new variable in BA2_AddInfection on the entity called BA2Infection_FirstTick, and set it to true.
- In BA2_InfectionTick, check for this variable. If it's true, we set the damage type for that infection tick to DMG_NERVEGAS. We then do the ent:Health == lastHealth check as normal, and set BA2Infection_FirstTick to false.

The first tick stuff is necessary because we don't want to cure the player when they enter a hazmat suit or something.