CURE SPRAY!
"It turns out that the Bio-Virus is far less intimidating if you can just magic-wand it away. If you find infection too annoying, you can still disable it for players and/or NPCs." --Sninctbur

so uhhh yeah about that one homie

Much of the code here is borrowed/adapted from BA1, with the model being the same also (minus the recompile I had to do).