I FOUND THE SOURCE OF THE LAG (and other changes!)
If there were a lot of zombies in a confined area without a target to attack, they used to lag the ever-loving shit out of the game. (This was an issue in the main addon too.) I found out why this was: computing a bad path was extremely expensive, and it was happening extremely often in confined areas because the zombie wandering system was putting the navigation target inside of the walls. The solution to this was in two steps:

1. Redo the wandering system. Instead of blindly selecting a random spot within 200 units, pathtrace to a random spot within 200 units. Navigate to wherever that pathtrace hits.

2. Bad paths are still gonna happen, so we need a way to handle them that isn't "cripple the game by repeatedly attempting to calculate a route that won't work."  With the new bad path handling, only successful paths will be used for navigation, and HandleStuck will be called in the case of a bad path.

These two things have been given their own disabled-by-default convars, as they're still pretty experimental.

OTHER CHANGES:

Deleted area branching. In retrospect, it kind of sucked. The radius system does the job of proximity spawning just fine.

Fixed a bug with the horde spawner where removed NPCs would remain in the internal entity list table.

Made some light changes to the experimental sections of the settings menus.

Removed ba2_hs_intervalperzombie. It really didn't have a reason to exist, and messing with it caused more issues than it solved.